:package: build: publish npm package
