+++ src/commands/index.ts
@@ -1,0 +1,2 @@
+export * from './exportFirestore'
+export * from './generate'
