+++ src/constants.ts
@@ -1,9 +1,15 @@
 const TOOL_NAME = 'gqlfake'
+const TOOL_DESCRIPTION = 'A CLI tool that can be used to create fake, structured data using GraphQL schemas to specify fields and data types'
+const VERSION = '0.1.5'
+
 const GENERATE_DIRECTIVE_NAME = 'generate'
 const GENERATE_DIRECTIVE_ARGUMENT_NAME = 'data'
 
 export {
     TOOL_NAME,
+    TOOL_DESCRIPTION,
+
+    VERSION,
     GENERATE_DIRECTIVE_NAME,
     GENERATE_DIRECTIVE_ARGUMENT_NAME
 }
